fixed the import error, deleted the unused files in FastFCN package, created notebook main
