--- models/FastFCN/encoding/datasets/__init__.py
@@ -1,19 +1,0 @@
-from .base import *
-from .coco import COCOSegmentation
-from .ade20k import ADE20KSegmentation
-from .pascal_voc import VOCSegmentation
-from .pascal_aug import VOCAugSegmentation
-from .pcontext import ContextSegmentation
-from .cityscapes import CitySegmentation
-
-datasets = {
-    'coco': COCOSegmentation,
-    'ade20k': ADE20KSegmentation,
-    'pascal_voc': VOCSegmentation,
-    'pascal_aug': VOCAugSegmentation,
-    'pcontext': ContextSegmentation,
-    'citys': CitySegmentation,
-}
-
-def get_segmentation_dataset(name, **kwargs):
-    return datasets[name.lower()](**kwargs)
