--- models/FastFCN/encoding/dilated/__init__.py
@@ -1,2 +1,0 @@
-"""Dilated ResNet and DenseNet"""
-from .resnet import *
