--- models/FastFCN/encoding/functions/__init__.py
@@ -1,3 +1,0 @@
-"""Encoding Autograd Fuctions"""
-from .syncbn import *
-from .encoding import *
