--- models/FastFCN/encoding/models/__init__.py
@@ -1,18 +1,0 @@
-from .model_zoo import get_model
-from .model_store import get_model_file
-from .base import *
-from .fcn import *
-from .psp import *
-from .encnet import *
-from .deeplabv3 import *
-
-
-def get_segmentation_model(name, **kwargs):
-    from .fcn import get_fcn
-    models = {
-        'fcn': get_fcn,
-        'psp': get_psp,
-        'encnet': get_encnet,
-        'deeplab': get_deeplab
-    }
-    return models[name.lower()](**kwargs)
