--- models/FastFCN/encoding/utils/__init__.py
@@ -1,19 +1,0 @@
-##+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-## Created by: Hang Zhang
-## ECE Department, Rutgers University
-## Email: zhang.hang@rutgers.edu
-## Copyright (c) 2017
-##
-## This source code is licensed under the MIT-style license found in the
-## LICENSE file in the root directory of this source tree
-##+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-
-"""Encoding Util Tools"""
-from .lr_scheduler import LR_Scheduler
-from .metrics import SegmentationMetric, batch_intersection_union, batch_pix_accuracy
-from .pallete import get_mask_pallete
-from .files import *
-
-__all__ = ['LR_Scheduler', 'batch_pix_accuracy', 'batch_intersection_union',
-           'save_checkpoint', 'download', 'mkdir', 'check_sha1',
-           'get_mask_pallete']
