+++ models/FastFCN/models/base.py
@@ -1,0 +1,74 @@
+###########################################################################
+# Created by: Hang Zhang 
+# Email: zhang.hang@rutgers.edu 
+# Copyright (c) 2017
+###########################################################################
+
+import math
+import numpy as np
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+import torchvision.models as models
+
+from .jpu import JPU, JPU_X
+
+up_kwargs = {'mode': 'bilinear', 'align_corners': True}
+
+__all__ = ['BaseNet', 'MultiEvalModule']
+
+class BaseNet(nn.Module):
+    def __init__(self, nclass, backbone, aux, se_loss, jpu="JPU", dilated=False, norm_layer=None,
+                 base_size=520, crop_size=480, mean=[.485, .456, .406],
+                 std=[.229, .224, .225], root='~/.encoding/models', **kwargs):
+        super(BaseNet, self).__init__()
+        self.nclass = nclass
+        self.aux = aux
+        self.se_loss = se_loss
+        self.mean = mean
+        self.std = std
+        self.base_size = base_size
+        self.crop_size = crop_size
+        # copying modules from pretrained models
+        if backbone == 'resnet50':
+            self.pretrained = models.resnet50(weights=models.ResNet50_Weights.IMAGENET1K_V2)
+        elif backbone == 'resnet101':
+            self.pretrained = models.resnet101(weights=models.ResNet101_Weights.IMAGENET1K_V2)
+        elif backbone == 'resnet152':
+            self.pretrained = models.resnet152(weights=models.ResNet152_Weights.IMAGENET1K_V2)
+        else:
+            raise RuntimeError('unknown backbone: {}'.format(backbone))
+        
+        # bilinear upsample options
+        self._up_kwargs = up_kwargs
+        self.backbone = backbone
+        self.jpu = None
+        if jpu == 'JPU':
+            self.jpu = JPU([512, 1024, 2048], width=512, norm_layer=norm_layer, up_kwargs=up_kwargs)
+        elif jpu == 'JPU_X':
+            self.jpu = JPU_X([512, 1024, 2048], width=512, norm_layer=norm_layer, up_kwargs=up_kwargs)
+
+    def base_forward(self, x):
+        x = self.pretrained.conv1(x)
+        x = self.pretrained.bn1(x)
+        x = self.pretrained.relu(x)
+        x = self.pretrained.maxpool(x)
+        c1 = self.pretrained.layer1(x)
+        c2 = self.pretrained.layer2(c1)
+        c3 = self.pretrained.layer3(c2)
+        c4 = self.pretrained.layer4(c3)
+
+        if self.jpu:
+            return self.jpu(c1, c2, c3, c4)
+        else:
+            return c1, c2, c3, c4
+
+    # def evaluate(self, x, target=None):
+    #     pred = self.forward(x)
+    #     if isinstance(pred, (tuple, list)):
+    #         pred = pred[0]
+    #     if target is None:
+    #         return pred
+    #     correct, labeled = batch_pix_accuracy(pred.data, target.data)
+    #     inter, union = batch_intersection_union(pred.data, target.data, self.nclass)
+    #     return correct, labeled, inter, union
